+++ src/UI/InfoDust/InfoDust.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { faInfoCircle } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Tooltip } from 'react-tooltip';
 import 'react-tooltip/dist/react-tooltip.css';
 import { MIN_DUST, FormDataTestIdsEnum } from 'constants/index';
 import { FormatAmount } from 'UI/FormatAmount';
+import { MvxCircleInfoIcon } from 'UI/sdkDappUi';
 import styles from './styles.module.scss';
 
 export const InfoDust = ({ egldLabel }: { egldLabel: string }) => (
@@ -30,7 +29,7 @@ export const InfoDust = ({ egldLabel }: { egldLabel: string }) => (
       data-testid={FormDataTestIdsEnum.infoDust}
       className={styles.infoDustTrigger}
     >
-      <FontAwesomeIcon icon={faInfoCircle} className='i-icon' />
+      <MvxCircleInfoIcon class={styles.infoDustIcon} />
     </div>
   </div>
 );